+++ src/main/java/com/braintreegateway/PaymentMethodRequest.java
@@ -18,6 +18,7 @@ public class PaymentMethodRequest extends Request {
     private String number;
     private String paymentMethodToken;
     private String paypalRefreshToken;
+    // NEXT_MAJOR_VERSION remove this variable as it has been ignored in the gateway
     private Boolean paypalVaultWithoutUpgrade;
     private String venmoSdkPaymentMethodCode;
 
@@ -74,6 +75,8 @@ public class PaymentMethodRequest extends Request {
         return this;
     }
 
+    // NEXT_MAJOR_VERSION remove this method as it has been ignored in the gateway
+    @Deprecated
     public PaymentMethodRequest paypalVaultWithoutUpgrade(Boolean paypalVaultWithoutUpgrade) {
         this.paypalVaultWithoutUpgrade = paypalVaultWithoutUpgrade;
         return this;
@@ -152,6 +155,7 @@ public class PaymentMethodRequest extends Request {
             addElement("expirationYear", expirationYear).
             addElement("paymentMethodNonce", paymentMethodNonce).
             addElement("paypalRefreshToken", paypalRefreshToken).
+            // NEXT_MAJOR_VERSION remove this element as it has been ignored in the gateway
             addElement("paypalVaultWithoutUpgrade", paypalVaultWithoutUpgrade).
             addElement("venmoSdkPaymentMethodCode", venmoSdkPaymentMethodCode);